Support nested braces autocomplete and BACKSPACE

+++ PowerShellTools.Test/IntelliSense/AutoCompletionControllerUnitTests.cs
@@ -26,7 +26,7 @@ namespace PowerShellTools.Test.IntelliSense
             _mockedScript = @"MyFunc";
             _mockedCaret = 6;
 
-            int actual = autoCompletionController.ProcessKeystroke((uint)VSConstants.VSStd2KCmdID.TYPECHAR, '(');
+            int actual = autoCompletionController.ProcessKeystroke(VSConstants.VSStd2KCmdID.TYPECHAR, '(');
             Assert.AreEqual<int>(VSConstants.S_OK, actual);
             Assert.AreEqual<string>(@"MyFunc()", _mockedScript);
             Assert.AreEqual<int>(7, _mockedCaret);
@@ -41,7 +41,7 @@ namespace PowerShellTools.Test.IntelliSense
             _mockedScript = @"MyFunc";
             _mockedCaret = 6;
 
-            int actual = autoCompletionController.ProcessKeystroke((uint)VSConstants.VSStd2KCmdID.TYPECHAR, '\"');
+            int actual = autoCompletionController.ProcessKeystroke(VSConstants.VSStd2KCmdID.TYPECHAR, '\"');
             Assert.AreEqual<int>(VSConstants.S_OK, actual);
             Assert.AreEqual<string>(@"MyFunc""""", _mockedScript);
             Assert.AreEqual<int>(7, _mockedCaret);
@@ -56,7 +56,7 @@ namespace PowerShellTools.Test.IntelliSense
             _mockedScript = @"MyFunc[]";
             _mockedCaret = 7;
 
-            int actual = autoCompletionController.ProcessKeystroke((uint)VSConstants.VSStd2KCmdID.TYPECHAR, ']');
+            int actual = autoCompletionController.ProcessKeystroke(VSConstants.VSStd2KCmdID.TYPECHAR, ']');
             Assert.AreEqual<int>(VSConstants.S_OK, actual);
             Assert.AreEqual<string>(@"MyFunc[]", _mockedScript);
             Assert.AreEqual<int>(_mockedCaret, 8);
@@ -71,7 +71,7 @@ namespace PowerShellTools.Test.IntelliSense
             _mockedScript = @"MyFunc""""";
             _mockedCaret = 7;
 
-            int actual = autoCompletionController.ProcessKeystroke((uint)VSConstants.VSStd2KCmdID.TYPECHAR, '\"');
+            int actual = autoCompletionController.ProcessKeystroke(VSConstants.VSStd2KCmdID.TYPECHAR, '\"');
             Assert.AreEqual<int>(VSConstants.S_OK, actual);
             Assert.AreEqual<string>(@"MyFunc""""", _mockedScript);
             Assert.AreEqual<int>(_mockedCaret, 8);
@@ -86,16 +86,16 @@ namespace PowerShellTools.Test.IntelliSense
             _mockedScript = @"MyFunc""""";
             _mockedCaret = 7;
 
-            int actual = autoCompletionController.ProcessKeystroke((uint)VSConstants.VSStd2KCmdID.TYPECHAR, 'a');
+            int actual = autoCompletionController.ProcessKeystroke(VSConstants.VSStd2KCmdID.TYPECHAR, 'a');
             Assert.AreEqual<int>(VSConstants.S_FALSE, actual);
 
-            actual = autoCompletionController.ProcessKeystroke((uint)VSConstants.VSStd2KCmdID.TYPECHAR, 'b');
+            actual = autoCompletionController.ProcessKeystroke(VSConstants.VSStd2KCmdID.TYPECHAR, 'b');
             Assert.AreEqual<int>(VSConstants.S_FALSE, actual);
 
-            actual = autoCompletionController.ProcessKeystroke((uint)VSConstants.VSStd2KCmdID.TYPECHAR, 'c');
+            actual = autoCompletionController.ProcessKeystroke(VSConstants.VSStd2KCmdID.TYPECHAR, 'c');
             Assert.AreEqual<int>(VSConstants.S_FALSE, actual);
 
-            actual = autoCompletionController.ProcessKeystroke((uint)VSConstants.VSStd2KCmdID.TYPECHAR, '\"');
+            actual = autoCompletionController.ProcessKeystroke(VSConstants.VSStd2KCmdID.TYPECHAR, '\"');
             Assert.AreEqual<int>(VSConstants.S_OK, actual);
         }
 
@@ -108,7 +108,7 @@ namespace PowerShellTools.Test.IntelliSense
             _mockedScript = @"MyFunc()";
             _mockedCaret = 7;
 
-            int actual = autoCompletionController.ProcessKeystroke((uint)VSConstants.VSStd2KCmdID.TYPECHAR, '\"');
+            int actual = autoCompletionController.ProcessKeystroke(VSConstants.VSStd2KCmdID.TYPECHAR, '\"');
             Assert.AreEqual<int>(VSConstants.S_OK, actual);
             Assert.AreEqual<string>(@"MyFunc("""")", _mockedScript);
             Assert.AreEqual<int>(_mockedCaret, 8);
@@ -123,7 +123,7 @@ namespace PowerShellTools.Test.IntelliSense
             _mockedScript = @"MyFunc()";
             _mockedCaret = 7;
 
-            int actual = autoCompletionController.ProcessKeystroke((uint)VSConstants.VSStd2KCmdID.TYPECHAR, ')');
+            int actual = autoCompletionController.ProcessKeystroke(VSConstants.VSStd2KCmdID.TYPECHAR, ')');
             Assert.AreEqual<string>(@"MyFunc()", _mockedScript);
             Assert.AreEqual<int>(7, _mockedCaret);
             Assert.AreEqual<int>(VSConstants.S_FALSE, actual);
@@ -137,7 +137,7 @@ namespace PowerShellTools.Test.IntelliSense
 
             _mockedScript = @"MyFunc()";
             _mockedCaret = 7;
-            int actual = autoCompletionController.ProcessKeystroke((uint)VSConstants.VSStd2KCmdID.RETURN);
+            int actual = autoCompletionController.ProcessKeystroke(VSConstants.VSStd2KCmdID.RETURN);
             Assert.AreEqual<int>(VSConstants.S_FALSE, actual);
         }
 
@@ -149,7 +149,7 @@ namespace PowerShellTools.Test.IntelliSense
 
             _mockedScript = @"MyFunc{}";
             _mockedCaret = 7;
-            int actual = autoCompletionController.ProcessKeystroke((uint)VSConstants.VSStd2KCmdID.RETURN);
+            int actual = autoCompletionController.ProcessKeystroke(VSConstants.VSStd2KCmdID.RETURN);
             Assert.AreEqual<int>(VSConstants.S_OK, actual);
             Assert.AreEqual<string>("MyFunc{\r\n\r\n}", _mockedScript);
             Assert.AreEqual<int>(9, _mockedCaret);
@@ -163,7 +163,7 @@ namespace PowerShellTools.Test.IntelliSense
 
             _mockedScript = @"MyFunc{}";
             _mockedCaret = 7;
-            int actual = autoCompletionController.ProcessKeystroke((uint)VSConstants.VSStd2KCmdID.RETURN);
+            int actual = autoCompletionController.ProcessKeystroke(VSConstants.VSStd2KCmdID.RETURN);
             Assert.AreEqual<int>(VSConstants.S_FALSE, actual);
             Assert.AreEqual<string>(@"MyFunc{}", _mockedScript);
             Assert.AreEqual<int>(7, _mockedCaret);
@@ -177,7 +177,7 @@ namespace PowerShellTools.Test.IntelliSense
 
             _mockedScript = @"MyFunc[]";
             _mockedCaret = 7;
-            int actual = autoCompletionController.ProcessKeystroke((uint)VSConstants.VSStd2KCmdID.RETURN);
+            int actual = autoCompletionController.ProcessKeystroke(VSConstants.VSStd2KCmdID.RETURN);
             Assert.AreEqual<int>(VSConstants.S_FALSE, actual);
             Assert.AreEqual<string>(@"MyFunc[]", _mockedScript);
             Assert.AreEqual<int>(7, _mockedCaret);
@@ -191,7 +191,7 @@ namespace PowerShellTools.Test.IntelliSense
 
             _mockedScript = @"MyFunc()";
             _mockedCaret = 7;
-            int actual = autoCompletionController.ProcessKeystroke((uint)VSConstants.VSStd2KCmdID.BACKSPACE);
+            int actual = autoCompletionController.ProcessKeystroke(VSConstants.VSStd2KCmdID.BACKSPACE);
             Assert.AreEqual<int>(VSConstants.S_FALSE, actual);
             Assert.AreEqual<string>(@"MyFunc()", _mockedScript);
             Assert.AreEqual<int>(7, _mockedCaret);
@@ -205,7 +205,7 @@ namespace PowerShellTools.Test.IntelliSense
 
             _mockedScript = @"MyFunc()";
             _mockedCaret = 7;
-            int actual = autoCompletionController.ProcessKeystroke((uint)VSConstants.VSStd2KCmdID.BACKSPACE);
+            int actual = autoCompletionController.ProcessKeystroke(VSConstants.VSStd2KCmdID.BACKSPACE);
             Assert.AreEqual<int>(VSConstants.S_FALSE, actual);
             Assert.AreEqual<string>(@"MyFunc()", _mockedScript);
             Assert.AreEqual<int>(7, _mockedCaret);
@@ -219,7 +219,7 @@ namespace PowerShellTools.Test.IntelliSense
 
             _mockedScript = @"MyFunc(";
             _mockedCaret = 7;
-            int actual = autoCompletionController.ProcessKeystroke((uint)VSConstants.VSStd2KCmdID.BACKSPACE);
+            int actual = autoCompletionController.ProcessKeystroke(VSConstants.VSStd2KCmdID.BACKSPACE);
             Assert.AreEqual<int>(VSConstants.S_FALSE, actual);
             Assert.AreEqual<string>(@"MyFunc(", _mockedScript);
             Assert.AreEqual<int>(7, _mockedCaret);
@@ -233,7 +233,7 @@ namespace PowerShellTools.Test.IntelliSense
 
             _mockedScript = @"MyFunc()";
             _mockedCaret = 7;
-            int actual = autoCompletionController.ProcessKeystroke((uint)VSConstants.VSStd2KCmdID.BACKSPACE);
+            int actual = autoCompletionController.ProcessKeystroke(VSConstants.VSStd2KCmdID.BACKSPACE);
             Assert.AreEqual<int>(VSConstants.S_OK, actual);
             Assert.AreEqual<string>(@"MyFunc", _mockedScript);
             Assert.AreEqual<int>(6, _mockedCaret);
@@ -247,7 +247,7 @@ namespace PowerShellTools.Test.IntelliSense
 
             _mockedScript = @"MyFunc""""";
             _mockedCaret = 7;
-            int actual = autoCompletionController.ProcessKeystroke((uint)VSConstants.VSStd2KCmdID.BACKSPACE);
+            int actual = autoCompletionController.ProcessKeystroke(VSConstants.VSStd2KCmdID.BACKSPACE);
             Assert.AreEqual<int>(VSConstants.S_OK, actual);
             Assert.AreEqual<string>(@"MyFunc", _mockedScript);
             Assert.AreEqual<int>(6, _mockedCaret);
